created files flagging and retrieveFunctions changed details to include these files
